moved code from findroots.py into claculus.py and deleted fendroots.py
